Getting range of dates needs a fix
